+++ internal/setup/cmd.go
@@ -1,0 +1,1 @@
+package setup
