+++ internal/setup/data.go
@@ -1,0 +1,65 @@
+package setup
+
+import (
+	"os"
+
+	"github.com/git-town/git-town/v21/internal/cli/dialog/dialogcomponents"
+	"github.com/git-town/git-town/v21/internal/cli/dialog/dialogdomain"
+	"github.com/git-town/git-town/v21/internal/config/cliconfig"
+	"github.com/git-town/git-town/v21/internal/config/gitconfig"
+	"github.com/git-town/git-town/v21/internal/execute"
+	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
+	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	"github.com/git-town/git-town/v21/internal/subshell/subshelldomain"
+	. "github.com/git-town/git-town/v21/pkg/prelude"
+)
+
+type SetupData struct {
+	backend       subshelldomain.Querier
+	dialogInputs  dialogcomponents.TestInputs
+	localBranches gitdomain.BranchInfos
+	remotes       gitdomain.Remotes
+}
+
+func LoadSetupData(repo execute.OpenRepoResult, cliConfig cliconfig.CliConfig) (data SetupData, exit dialogdomain.Exit, err error) {
+	dialogTestInputs := dialogcomponents.LoadTestInputs(os.Environ())
+	repoStatus, err := repo.Git.RepoStatus(repo.Backend)
+	if err != nil {
+		return data, false, err
+	}
+	branchesSnapshot, _, _, exit, err := execute.LoadRepoSnapshot(execute.LoadRepoSnapshotArgs{
+		Backend:               repo.Backend,
+		CommandsCounter:       repo.CommandsCounter,
+		ConfigSnapshot:        repo.ConfigSnapshot,
+		Connector:             None[forgedomain.Connector](),
+		Detached:              false,
+		DialogTestInputs:      dialogTestInputs,
+		Fetch:                 false,
+		FinalMessages:         repo.FinalMessages,
+		Frontend:              repo.Frontend,
+		Git:                   repo.Git,
+		HandleUnfinishedState: true,
+		Repo:                  repo,
+		RepoStatus:            repoStatus,
+		RootDir:               repo.RootDir,
+		UnvalidatedConfig:     repo.UnvalidatedConfig,
+		ValidateNoOpenChanges: false,
+		Verbose:               cliConfig.Verbose,
+	})
+	if err != nil {
+		return data, exit, err
+	}
+	remotes, err := repo.Git.Remotes(repo.Backend)
+	if err != nil {
+		return data, exit, err
+	}
+	if len(remotes) == 0 {
+		remotes = gitdomain.Remotes{gitconfig.DefaultRemote(repo.Backend)}
+	}
+	return SetupData{
+		backend:       repo.Backend,
+		dialogInputs:  dialogTestInputs,
+		localBranches: branchesSnapshot.Branches,
+		remotes:       remotes,
+	}, exit, nil
+}
