+++ internal/setup/doc.go
@@ -1,0 +1,2 @@
+// Package setup implements the Git Town setup assistant.
+package setup
